feat: 스케줄러를 통한 벌크 인서트 - 주기적으로 벌크 인서트를 통해 정합성 관리

+++ src/main/java/com/naver/playlist/domain/service/PlayListAsyncService.java
@@ -5,8 +5,11 @@ import com.naver.playlist.domain.redis.RedisHashService;
 import com.naver.playlist.domain.repository.JdbcBulkRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import static com.naver.playlist.domain.constant.RedisConstants.PLAY_LIST_HASH_NAME;
 
 @Service
 @RequiredArgsConstructor
@@ -18,6 +21,19 @@ public class PlayListAsyncService {
     @Async
     @Transactional
     public void bulkInsertAsync(PlayListCreateResponse dto) {
+        bulkRepository.bulkInsert(dto.getPlayList());
+        redisHashService.deleteSnapshot(dto.getKey());
+    }
+
+    @Scheduled(fixedRate = 1000)
+    @Transactional
+    public void syncPlayList() {
+        PlayListCreateResponse dto = redisHashService.extractPlayList(PLAY_LIST_HASH_NAME);
+
+        if (dto == null) {
+            return;
+        }
+
         bulkRepository.bulkInsert(dto.getPlayList());
         redisHashService.deleteSnapshot(dto.getKey());
     }
